requirements.txt
alembic==1.16.1
aniso8601==10.0.1
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
python-decouple==3.4
Flask==3.1.1
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
pytz==2025.2
resources==0.0.1
six==1.17.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.13.2
utils==1.0.2
Werkzeug==3.1.3
Flask-RESTful
Flask-Migrate
Flask-SQLAlchemy
python-decouple
pymysql
